# Cache `dir(Module)`.